Factor out AudienceAndMusicians hooks
